+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Made EuiIcon a PureComponent, to speed up React re-render performance ([#2448](https://github.com/elastic/eui/pull/2448))
 - Added ability for `EuiColorStops` to accept user-defined range bounds ([#2396](https://github.com/elastic/eui/pull/2396))
+- Added `external` prop to `EuiLink` ([#2442](https://github.com/elastic/eui/pull/2442))
 
 ## [`14.6.0`](https://github.com/elastic/eui/tree/v14.6.0)